+++ NEWS.md
@@ -1,3 +1,8 @@
+# surveyexplorer 0.2.0
+
+-   Add ability to specify column order with matrix_table function
+-   Set default order for multi_table for results to be shown from high to low remove frequency summary row for non-grouped data
+
 # surveyexplorer 0.1.0
 
 -   Initial submission to CRAN